add kruskal and prim algorithms
* add isPending

* add isRegular

* add isNull

* add isNull and isSimple

* add test isSimple

* add isComplete

* add isConnected without graph start at 0 verify

* refactor all tests still without isConnected 0

* add comments

* refactor isConnected

* add adjmatrix

* add prims without rules

* start kruskal

* add kruskal

* minor changes

* add comments

* fix kruskal

* fix after merge

+++ src/main/java/models/DirectedEdge.java
@@ -12,7 +12,7 @@ public class DirectedEdge extends Edge {
      * @param fileString String from file
      */
     DirectedEdge(String fileString) {
-        super(Double.parseDouble(fileString.split(";")[2]));
+        super(Integer.parseInt(fileString.split(";")[2]));
 
         String[] strings = fileString.split(";");
         if (strings[3].equals("1")) {
@@ -31,7 +31,7 @@ public class DirectedEdge extends Edge {
      * @param v1     first vertice
      * @param v2     second vertice
      */
-    DirectedEdge(double weight, int v1, int v2) {
+    DirectedEdge(int weight, int v1, int v2) {
         super(weight);
         this.vertices.add(new Vertice(v1));
         this.vertices.add(new Vertice(v2));
